rename value module
